Refactor AbstractClipboard and Clipboard into own file
Removed from interface.py
